Added some more 64bit warning fixes.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@23616 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/sdl/archdep_win32.c
@@ -185,7 +185,7 @@ char *archdep_program_name(void)
         if (e == NULL) {
             e = argv0 + strlen(argv0);
         }
-        len = e - s + 1;
+        len = (int)(e - s + 1);
         program_name = lib_malloc(len);
         memcpy(program_name, s, len - 1);
         program_name[len - 1] = 0;
@@ -371,7 +371,7 @@ int archdep_spawn(const char *name, char **argv, char **pstdout_redir, const cha
     }
 
     /* Spawn the child process.  */
-    retval = _spawnvp(_P_WAIT, name, (const char **)argv);
+    retval = (int)_spawnvp(_P_WAIT, name, (const char **)argv);
 
 cleanup:
     if (old_stdout >= 0) {
